+++ MyApplication/app/src/main/java/com/example/myapplication/data/dao/BudgetDAO.java
@@ -46,7 +46,7 @@ public interface BudgetDAO {
     LiveData<List<BudgetItem>>getBudgetItems(int userId, int month, int yearInt, String monthText, String year);
 
     @Query("UPDATE budgets SET categoryId = :categoryId, limitAmount = :limitAmount, updatedAt = :updatedAt WHERE id = :budgetId")
-    int updateBudget(int budgetId, int categoryId, int limitAmount, long updatedAt);
+    int updateBudget(int budgetId, int categoryId, long limitAmount, long updatedAt);
 
     @Query("DELETE FROM budgets WHERE id = :budgetId")
     int deleteById(int budgetId);
